% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R,
%   R/plot.multiple.sub.seasonal.series.R
\name{accuracy.of.specific.horizen.cal}
\alias{accuracy.of.specific.horizen.cal}
\title{calculate accuracy of specific horizen,such as Monthly data, accuracy of horizen 7-12}
\usage{
\method{accuracy}{of.specific.horizen.cal}(accuracy.of.every.horizen.m, begin.horizen, end.horizen)

\method{accuracy}{of.specific.horizen.cal}(accuracy.of.every.horizen.m, begin.horizen, end.horizen)
}
\arguments{
\item{accuracy.of.every.horizen.m}{shape(num.of.ts*all.horizen)}

\item{begin.position}{such as 7}

\item{end.position}{such as 12}

\item{level.value}{interval prediction, such as 50, 85, 95%}
}
\value{
forecasts which contain point and interval forecasts

forecasts which contain point and interval forecasts
}
\description{
calculate accuracy of specific horizen,such as Monthly data, accuracy of horizen 7-12

calculate accuracy of specific horizen,such as Monthly data, accuracy of horizen 7-12
}
\examples{
# Not Run 
calcualte mean
# Not Run 
}
