% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{accuy_of_specific_horizen_cal}
\alias{accuy_of_specific_horizen_cal}
\title{calculate accuracy of specific horizen,such as Monthly data, accuracy of horizen 7-12
test}
\usage{
accuy_of_specific_horizen_cal(
  accuracy.of.every.horizen.m,
  begin.horizen,
  end.horizen
)
}
\arguments{
\item{accuracy.of.every.horizen.m}{shape(num.of.ts,all.horizen)}

\item{begin.position}{such as 7}

\item{end.position}{such as 12}
}
\value{
forecasts which contain point and interval forecasts
}
\description{
calculate accuracy of specific horizen,such as Monthly data, accuracy of horizen 7-12
test
}
