% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{baseline.method}
\alias{baseline.method}
\title{baseline method for one ts}
\usage{
baseline.method(ts, method.option, level.value)
}
\arguments{
\item{ts}{time series}

\item{method.option}{'ets' or 'arima'}

\item{level.value}{95, 90, 85}
}
\value{
calculate forecasts of baseline
}
\description{
baseline method for one ts
}
