% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{batch_data_baseline_method}
\alias{batch_data_baseline_method}
\title{forecast for many ts using benchmark method with parallel computing}
\usage{
batch_data_baseline_method(
  dataset,
  data.index,
  ts.type,
  horizen,
  method.option,
  level.value
)
}
\arguments{
\item{dataset}{eg. M4}

\item{data.index}{eg. 1:100}

\item{ts.type}{'M' or 'Q'}

\item{horizen}{h}

\item{method.option}{'ets' or 'arima'}

\item{level.value}{95}
}
\value{
Null
}
\description{
forecast for many ts using benchmark method with parallel computing
}
