% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{combined_date_formatting}
\alias{combined_date_formatting}
\title{formate date}
\usage{
combined_date_formatting(original.date, type.of.ts)
}
\arguments{
\item{original.date}{original date}

\item{type.of.ts}{'M' or 'Q'}
}
\value{
formatted date
}
\description{
formate date
}
