% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R,
%   R/plot.multiple.sub.seasonal.series.R
\name{forecast.method}
\alias{forecast.method}
\title{get point,interval forecasts of one method}
\usage{
\method{forecast}{method}(method.option, x, h, level.value)

\method{forecast}{method}(method.option, x, h, level.value)
}
\arguments{
\item{method.option}{method name, such as 'ets' or'arima'}

\item{x}{historical data of ts, such as M3[[100]]$x}

\item{h}{forecasting horizen}

\item{level.value}{interval prediction, such as 50, 85, 95%}
}
\value{
forecasts which contain point and interval forecasts

forecasts which contain point and interval forecasts
}
\description{
get point,interval forecasts of one method

get point,interval forecasts of one method
}
\examples{
# Not Run 
# Not Run 
}
