% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{forecasting_with_mutiple_approches_for_batch_data_and_saving_results}
\alias{forecasting_with_mutiple_approches_for_batch_data_and_saving_results}
\title{forecast with parellel computing for batch data(many time series)
This function allows you}
\usage{
forecasting_with_mutiple_approches_for_batch_data_and_saving_results(
  dataset,
  method.option,
  level.value,
  saving.path
)
}
\arguments{
\item{dataset}{eg. M4}

\item{method.option}{'ets' or 'arima'}

\item{level.value}{95}

\item{saving.path}{path of saving intermediate results}

\item{horizen}{}
}
\value{
Null
}
\description{
forecast with parellel computing for batch data(many time series)
This function allows you
}
