% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R,
%   R/plot.multiple.sub.seasonal.series.R
\name{format.month}
\alias{format.month}
\title{get formatted month of the original month}
\usage{
\method{format}{month}(original.month)

\method{format}{month}(original.month)
}
\arguments{
\item{original.month}{eg.c(11,12,13,14,15)}
}
\value{
eg.c(11,12,13,14,15)---->c(11,12,1,2,3)

eg.c(11,12,13,14,15)---->c(11,12,1,2,3)
}
\description{
get formatted month of the original month

get formatted month of the original month
}
\examples{
# Not Run
# Not Run
}
