% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiple.sub.seasonal.series.R
\name{format.month1}
\alias{format.month1}
\title{get formatted month of the original month 
 eg.c(11,12,13,14,15)---->c(11,12,1,2,3)}
\usage{
\method{format}{month1}(original.month)
}
\arguments{
\item{original.month}{eg.c(11,12,13,14,15)}
}
\value{
eg.c(11,12,13,14,15)---->c(11,12,1,2,3)
}
\description{
get formatted month of the original month 
 eg.c(11,12,13,14,15)---->c(11,12,1,2,3)
}
