% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R,
%   R/plot.multiple.sub.seasonal.series.R
\name{format.quarter}
\alias{format.quarter}
\title{get formatted quartely of the original quarterly}
\usage{
\method{format}{quarter}(original.quarter)

\method{format}{quarter}(original.quarter)
}
\arguments{
\item{original.quarter}{eg. c(3,4,5,7)}
}
\value{
eg.c(3,4,5,7)---->c(3,4,1,3)

eg.c(3,4,5,7)---->c(3,4,1,3)
}
\description{
get formatted quartely of the original quarterly

get formatted quartely of the original quarterly
}
\examples{
# Not Run
# Not Run
}
