% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiple.sub.seasonal.series.R
\name{format_combined_date}
\alias{format_combined_date}
\title{format combined date1}
\usage{
format_combined_date(original.date, type.of.ts)
}
\arguments{
\item{original.date}{eg. c(3,4,5,7)}

\item{type.of.ts}{'M' or 'Q'}
}
\value{
eg.c(3,4,5,7)---->c(3,4,1,3)
}
\description{
format combined date1
}
