% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiple.sub.seasonal.series.R
\name{format_quarter}
\alias{format_quarter}
\title{get formatted quartely of the original quarterly}
\usage{
format_quarter(original.quarter)
}
\arguments{
\item{original.quarter}{eg. c(3,4,5,7)}
}
\value{
eg.c(3,4,5,7)---->c(3,4,1,3)
}
\description{
get formatted quartely of the original quarterly
}
