% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{get.corresponding.ts.data.according.to.the.combination.months}
\alias{get.corresponding.ts.data.according.to.the.combination.months}
\title{get the corresponding ts value according to the combinded months
test}
\usage{
get.corresponding.ts.data.according.to.the.combination.months(
  ts,
  combination.months,
  formatted.date.of.the.ts
)
}
\arguments{
\item{ts}{train data of the ts, eg.}

\item{combination.months}{combined month, eg,c(1,2,3)}

\item{formatted.date.of.the.ts}{formatted date for the train data of ts}
}
\value{
new.ts : new ts constructed from the orginal ts
}
\description{
get the corresponding ts value according to the combinded months
test
}
