% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiple.sub.seasonal.series.R
\name{get.corresponding.ts.data.according.to.the.combination.months1}
\alias{get.corresponding.ts.data.according.to.the.combination.months1}
\title{get the corresponding ts value according to the combinded months}
\usage{
get.corresponding.ts.data.according.to.the.combination.months1(
  ts,
  combination.months,
  formatted.date.of.the.ts
)
}
\arguments{
\item{ts}{train data of the ts, eg. M4[[1]]
@param combination.months  combined month, eg,c(1,2,3)
@param formatted.date.of.the.ts  formatted date for the train data of ts}
}
\value{
new.ts : new ts constructed from the orginal ts
}
\description{
get the corresponding ts value according to the combinded months
}
