% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiple.sub.seasonal.series.R
\name{get.different.sub.seasonal.seqence.for.one.ts1}
\alias{get.different.sub.seasonal.seqence.for.one.ts1}
\title{forecast every new ts}
\usage{
get.different.sub.seasonal.seqence.for.one.ts1(
  i,
  j,
  ts,
  formatted.date.of.the.train.data,
  formatted.date.of.the.test.data,
  type.of.ts
)
}
\arguments{
\item{i}{number of months (the number of sub-seasonal)
@param j  begin position}

\item{formatted.date.of.the.train.data}{formatted date of train data of ts}

\item{formatted.date.of.the.test.data}{formatted date of train data of ts}
}
\value{
forecasts of the new ts
}
\description{
forecast every new ts
}
