% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R,
%   R/plot.multiple.sub.seasonal.series.R
\name{get.the.formatted.date.of.the.train.data}
\alias{get.the.formatted.date.of.the.train.data}
\title{get formatted date of the train data
eg. #type.of.ts=='Quarterly', change month to quarterly format such as 2 3 4 1 2 3 4 1}
\usage{
get.the.formatted.date.of.the.train.data(ts, type.of.ts)

get.the.formatted.date.of.the.train.data(ts, type.of.ts)
}
\arguments{
\item{ts}{train data of time series eg. M4[[1]]}

\item{type.of.ts}{type of ts eg. 'M' or'Q'}
}
\value{
formatted.date.of.test.data eg. #type.of.ts=='Quarterly', change month to quarterly format such as 2 3 4 1 2 3 4 1

formatted.date.of.test.data eg. #type.of.ts=='Quarterly', change month to quarterly format such as 2 3 4 1 2 3 4 1
}
\description{
get formatted date of the train data
eg. #type.of.ts=='Quarterly', change month to quarterly format such as 2 3 4 1 2 3 4 1

get formatted date of the train data
eg. #type.of.ts=='Quarterly', change month to quarterly format such as 2 3 4 1 2 3 4 1
}
\examples{
# Not Run
# Not Run
}
