% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{mase_cal}
\alias{mase_cal}
\title{caculate mase
test}
\usage{
mase_cal(insample, outsample, forecasts)
}
\arguments{
\item{insample}{x}

\item{outsample}{xx}

\item{forecasts}{forecasts}
}
\value{
smape
}
\description{
caculate mase
test
}
