% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{method_forecast}
\alias{method_forecast}
\title{get point and interva forecasts of one method}
\usage{
method_forecast(method.option, x, h, level.value)
}
\arguments{
\item{method.option}{'ets' or 'arima'}

\item{x}{ts}

\item{h}{horizen}

\item{level.value}{95, 90,85}
}
\value{
A matrix of the infile
}
\description{
get point and interva forecasts of one method
}
