% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{month_formatting}
\alias{month_formatting}
\title{get formatted month of the original month 
 eg.c(11,12,13,14,15)---->c(11,12,1,2,3)}
\usage{
month_formatting(original.month)
}
\arguments{
\item{original.month}{eg.c(11,12,13,14,15)}
}
\value{
eg.c(11,12,13,14,15)---->c(11,12,1,2,3)
}
\description{
get formatted month of the original month 
 eg.c(11,12,13,14,15)---->c(11,12,1,2,3)
}
