% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{msis_cal}
\alias{msis_cal}
\title{caculate msis
test}
\usage{
msis_cal(insample, outsample, forecasts.lower, forecasts.upper, a)
}
\arguments{
\item{insample}{x}

\item{outsample}{xx}

\item{forecasts.lower}{lower forecasts}

\item{forecasts.upper}{upper forecasts}

\item{a}{0.05}
}
\value{
msis value
}
\description{
caculate msis
test
}
