% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{new_ts_forecast}
\alias{new_ts_forecast}
\title{forecast every new ts}
\usage{
new_ts_forecast(
  i,
  j,
  ts,
  formatted.date.of.the.train.data,
  formatted.date.of.the.test.data,
  method.option,
  level.value,
  type.of.ts
)
}
\arguments{
\item{i}{number of months (the number of sub-seasonal)
@param j  begin position}

\item{formatted.date.of.the.train.data}{formatted date of train data of ts}

\item{formatted.date.of.the.test.data}{formatted date of train data of ts}
}
\value{
forecasts of the new ts
}
\description{
forecast every new ts
}
