% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiple.sub.seasonal.series.R
\name{plot_different_sub_seasonal_series}
\alias{plot_different_sub_seasonal_series}
\title{plot different sub seasonal series}
\usage{
plot_different_sub_seasonal_series(ts, type.of.ts)
}
\arguments{
\item{ts}{time series,type: list}

\item{type.of.ts}{time series type 'M' or 'Q'}
}
\value{
plot an figure
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
