% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.plot.R
\name{plot_ets_components}
\alias{plot_ets_components}
\title{plot ets components updated version}
\usage{
plot_ets_components(ts.type, forecast.results)
}
\arguments{
\item{ts.type}{'M'or'Q'}

\item{forecast.results}{intermediate.results}
}
\value{
plot an figure
}
\description{
plot ets components updated version
}
