% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.plot.R
\name{plot_ets_components_for_batch_ts}
\alias{plot_ets_components_for_batch_ts}
\title{plot ets components for batch ts}
\usage{
plot_ets_components_for_batch_ts(
  ts.type,
  intermediate.results,
  dataset,
  saving.path
)
}
\arguments{
\item{ts.type}{'M' or 'Q'}

\item{intermediate.results}{intermediate.results}

\item{dataset}{M3, M1, or M4}

\item{saving.path}{path for saving figures}
}
\value{
plot an figure
}
\description{
plot ets components for batch ts
}
