% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.plot.R
\name{plot_ets_components_for_one_ts}
\alias{plot_ets_components_for_one_ts}
\title{plot ets components for one ts}
\usage{
plot_ets_components_for_one_ts(ts.index, ts.type, intermediate.results)
}
\arguments{
\item{ts.index}{time series index}

\item{ts.type}{'M'or'Q'}

\item{intermediate.results}{intermediate.results}
}
\value{
char
}
\description{
plot ets components for one ts
}
