% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.plot.R
\name{plot_forecasts_for_batch_ts}
\alias{plot_forecasts_for_batch_ts}
\title{plot forecasts for batch.ts}
\usage{
plot_forecasts_for_batch_ts(ts.type, intermediate.results, saving.path)
}
\arguments{
\item{ts.type}{'M' or 'Q'}

\item{intermediate.results}{intermediate.results}

\item{saving.path}{path for saving figures}
}
\value{
plot an figure
}
\description{
plot forecasts for batch.ts
}
