% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.plot.R
\name{plot_forecasts_of_new_mutiple_series}
\alias{plot_forecasts_of_new_mutiple_series}
\title{plot forecasts of new mutiple series}
\usage{
plot_forecasts_of_new_mutiple_series(ts.object, ts.type, forecasts.result)
}
\arguments{
\item{ts.object}{time series,type: list}

\item{ts.type}{time series type 'M' or 'Q'}

\item{forecasts.result}{list}
}
\value{
plot an figure
}
\description{
plot forecasts of new mutiple series
}
