% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.plot.R
\name{plot_model_components}
\alias{plot_model_components}
\title{plot model components}
\usage{
plot_model_components(ts.type, forecast.results, method.name)
}
\arguments{
\item{ts.type}{time series type 'M' or 'Q'}

\item{forecast.results}{forecasting result list}

\item{method.name}{'ets' or 'arima'}
}
\value{
A matrix of the infile
}
\description{
plot model components
}
