% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.plot.R
\name{plot_time_series_and_its_forecast}
\alias{plot_time_series_and_its_forecast}
\title{plot time series and its forecast}
\usage{
plot_time_series_and_its_forecast(
  training.value,
  testing.value,
  forecast.results
)
}
\arguments{
\item{training.value}{ts object}

\item{testing.value}{ts object}

\item{forecast.results}{list}
}
\value{
plot an figure
}
\description{
plot time series and its forecast
}
