% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.plot.R
\name{project.char.to.number.for.ets.method}
\alias{project.char.to.number.for.ets.method}
\title{project char to number for ets method}
\usage{
project.char.to.number.for.ets.method(components)
}
\arguments{
\item{components}{model components}
}
\value{
char
}
\description{
project char to number for ets method
}
