% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{quarter_formatting}
\alias{quarter_formatting}
\title{get formatted quartely of the original quarterly eg.c(3,4,5,7)---->c(3,4,1,3)}
\usage{
quarter_formatting(original.quarter)
}
\arguments{
\item{original.quarter}{eg. c(3,4,5,7)}
}
\value{
eg.c(3,4,5,7)---->c(3,4,1,3)
}
\description{
get formatted quartely of the original quarterly eg.c(3,4,5,7)---->c(3,4,1,3)
}
