% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.with.sub.seasons.R
\name{smape_cal}
\alias{smape_cal}
\title{caculate smape
test}
\usage{
smape_cal(outsample, forecasts)
}
\arguments{
\item{outsample}{x}

\item{forecasts}{forecasts}
}
\value{
smape
}
\description{
caculate smape
test
}
